Add omni training pipeline updates

+++ model/model_omni.py
@@ -346,7 +346,7 @@ class MiniMindOmni(MiniMindForCausalLM):
                 out = self.forward(torch.cat((audio_buffer[:, :, -1:], input_ids[:, -1:].unsqueeze(1)), dim=1), past_key_values=past_kvs, use_cache=use_cache, **args)
             past_kvs = out.past_key_values
 
-            logits = out.logits[0, -1, :].clone() / (temperature + 1e-9)
+            logits = out.logits[0, -1, :].clone().float() / (temperature + 1e-9)
             if rp != 1.0:
                 seen = list(set(input_ids[0].tolist())); score = logits[seen]; logits[seen] = torch.where(score > 0, score / rp, score * rp)
             if top_p and top_p < 1.0:
@@ -370,7 +370,7 @@ class MiniMindOmni(MiniMindForCausalLM):
                 if audio_step < i:
                     audio_codes[i].append(self.audio_pad_token)
                 else:
-                    logits_i = al[0, -1, :].clone() / 0.2
+                    logits_i = al[0, -1, :].clone().float() / 0.2
                     for prev_code in audio_codes[i][-3:]: score = logits_i[prev_code]; logits_i[prev_code] = torch.where(score > 0, score / 1.05, score * 1.05)
                     top_val, top_idx = logits_i.topk(50)
                     code = top_idx[torch.multinomial(F.softmax(top_val, dim=-1), 1)].item()
